requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
pycryptodome==3.19.0
networkx==3.2.1
requests